api/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = ""
authors = [
    {name = "GR",email = "rodrigg2@mail.uc.edu"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio (>=1.69.0,<2.0.0)",
    "grpcio-tools (>=1.69.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
